requirements.txt
weaviate-client==4.6.1
